Add rudimentary ability to respond to resizes
Long ago, in the days of the Incas, we were unable to properly
respond to certain window events. The window would expand and yet
the game itself would remain the same. Worse yet, it might respond
unpredictably. The mighty Inca warrior Josias, who was decidedly
not actually Inca, decided to defeat this foe. He changed the
program to respond to the appropriate window events, updating the
internal state of the program to reflect the current size of the
window. There is still much work to be done, but Josias will not
rest until the system becomes a cohesive whole. A good warrior
never tires.