Calculate position of board properly
Previously the game was confused. We taught the initial object to
behave appropriately, but it didn't pass on that vital knowledge
to its parent. Without that information, future offspring were
not able to benefit from this knowledge. Communication broke down
and the game displayed strange shapes that were considered
undesirable to most.

This has since been remedied, and now the whole family is at peace
again.